requirements.txt
streamlit
streamlit-audiorecorder
fastapi
uvicorn
python-multipart
numpy
pandas
librosa
openai-whisper
textblob
pyyaml
matplotlib
scipy
torch
ffmpeg-python
grafana
mlflow
prometheus-client